Move session stuff to a new module
